LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/border-image-width-expected.txt


PASS Can set 'border-image-width' to CSS-wide keywords: initial
PASS Can set 'border-image-width' to CSS-wide keywords: inherit
PASS Can set 'border-image-width' to CSS-wide keywords: unset
PASS Can set 'border-image-width' to CSS-wide keywords: revert
FAIL Can set 'border-image-width' to var() references:  var(--A) assert_equals: expected 2 but got 1
FAIL Can set 'border-image-width' to a length: 0px assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a length: -3.14em assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a length: 3.14cm assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a length: calc(0px + 0em) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'border-image-width' to a percent: 0% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a percent: -3.14% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a percent: 3.14% assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a percent: calc(0% + 0%) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'border-image-width' to a number: 0 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a number: -3.14 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a number: 3.14 assert_equals: expected "CSSUnitValue" but got "CSSStyleValue"
FAIL Can set 'border-image-width' to a number: calc(2 + 3) assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSStyleValue]"
FAIL Can set 'border-image-width' to the 'auto' keyword: auto assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
PASS Setting 'border-image-width' to a time throws TypeError
PASS Setting 'border-image-width' to an angle throws TypeError
PASS Setting 'border-image-width' to a flexible length throws TypeError
PASS Setting 'border-image-width' to a URL throws TypeError
PASS Setting 'border-image-width' to a transform throws TypeError
PASS 'border-image-width' does not supported '2em 3em'
PASS 'border-image-width' does not supported '5% 15% 10%'
PASS 'border-image-width' does not supported '5% 2em 10% auto'
